Got StompJS client to work with redux-saga; watchTheSocket generator fn can now dispatch timer updates
