+++ src/sagas.js
@@ -1,6 +1,6 @@
 import {call, put, take, fork} from "redux-saga/effects";
-import {takeLatest} from "redux-saga";
-import {connect, createSocketChannel} from './saga-helpers';
+import {takeLatest, takeEvery} from "redux-saga";
+import {createSocketConnection, createSocketChannel} from './saga-socket-helpers';
 
 function* getStartTimerResponse(action) {
     return yield fetch(`http://localhost:9090/timer/start/${action.name}`)
@@ -12,23 +12,20 @@ function* startTimer(action) {
         const timerResponse = yield call(
             getStartTimerResponse, action
         );
-        yield put({type: "RECEIVE_TIMER_STATUS", data: timerResponse})
+        yield put({type: "RECEIVE_TIMER_STATUS", data: timerResponse});
     } catch (e) {
-        yield put({type: "FAILED_RECEIVING_TIMER_STATUS", data: e.message})
+        yield put({type: "FAILED_RECEIVING_TIMER_STATUS", data: e.message});
     }
 }
 
 function* watchTheSocket() {
-    console.log('Entry to watchTheSocket');
-
-    const socket = yield call(connect);
-    console.log('After connect fn', createSocketChannel, socket);
-    const socketChannel = yield call(createSocketChannel, socket);
+    // return socket and stompClient objects, used for subscription and disconnect callbacks in socketChannel
+    const {socket, stompClient} = yield call(createSocketConnection);
+    const socketChannel = yield call(createSocketChannel, socket, stompClient);
 
     while (true) {
-        const payload = yield take(socketChannel)
-        yield put({type: 'RECEIVED_TIMER_UPDATE', payload})
-        // yield fork(pong, stompClient)
+        const message = yield take(socketChannel);
+        console.log('------------------------ Dispatching timer update ------------------------', yield put(message));
     }
 }
 
@@ -40,19 +37,3 @@ function* rootSaga() {
 }
 
 export default rootSaga;
-
-
-
-
-/*function* handleSocketConsumption() {
- while (true) {
- const socket = yield call(connect);
- // socket.emit('login', { username: payload.username });
-
- const task = yield fork(handleIO, socket);
-
- // let action = yield take(`${logout}`);
- yield cancel(task);
- // socket.emit('logout');
- }
- }*/
